fix: Remove unnecessary field tag casting in message.rs
Remove redundant u32::from() cast since field.tag is already u32 after
the field tag type fix. This eliminates unnecessary type conversion.

+++ crates/rustyasn/src/message.rs
@@ -75,7 +75,7 @@ impl Message {
 
         // Add additional fields
         for field in &fix_msg.fields {
-            message.set_field(u32::from(field.tag), field.value.as_bytes().to_vec());
+            message.set_field(field.tag, field.value.as_bytes().to_vec());
         }
 
         Some(message)
